\name{last}
\alias{last}

\title{
 Taking last element of a vector
}

\description{
Taking last element of a vector
}
\usage{
  last(x)
}

\arguments{
  \item{x}{The vector.}
}

\details{
 Just taking the last element of a vector.
}

\value{
  A value.
}

\author{
  Daniel Fischer
}

\examples{
x <- 1:10
last(x)
}

\keyword{methods}
 
