\name{Luke-package}
\alias{R/Luke-package}
\docType{package}

\title{
  Collection of useful Subfunctions.
}

\description{
  A loose collection of useful subfunctions
  
  https://github.com/fischuu/Luke
}

\details{
  \tabular{ll}{
    Package: \tab Luke\cr
    Type: \tab Package\cr
    Version: \tab 0.1.3\cr
    Date: \tab 2022-08-29\cr
    License: \tab GPL\cr
    LazyLoad: \tab yes\cr
  }
}

\author{
  Daniel Fischer

  Maintainer: Daniel Fischer <daniel.fischer@mtt.fi>
}

\keyword{multivariate}

