\name{lastElement}
\alias{lastElement}

\title{
 Taking last element of a vector
}

\description{
Taking last element of a vector
}
\usage{
  lastElement(x)
}

\arguments{
  \item{x}{The vector.}
}

\details{
 Just taking the last element of a vector.
}

\value{
  A value.
}

\author{
  Daniel Fischer
}

\examples{
x <- 1:10
lastElement(x)
}

\keyword{methods}
 
